backend/requirements.txt
uvicorn[standard]==0.29.0
fastapi==0.115.2
networkx==3.4.2
protobuf<6.0.0,>=3.20.2
pydantic==2.11.1
python-dotenv==1.0.1
requests==2.32.3
tree_sitter==0.24.0
tree_sitter_c==0.23.5
tree_sitter_cpp==0.23.4
tree_sitter_java==0.23.5
tree_sitter_javascript==0.23.1
tree_sitter_python==0.23.6
tree_sitter_typescript==0.23.2
google-ai-generativelanguage==0.6.15
google-genai==0.3.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pgvector==0.3.2
redis==5.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
slowapi==0.1.9
celery==5.4.0
pytest==7.4.3
httpx==0.25.2